requirements.txt
pymupdf
fastapi==0.109.2
uvicorn==0.27.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0
httpx==0.27.0
python-multipart
pillow>=10.4.0
sqlalchemy
asyncpg
google-generativeai